+++ src/main/java/com/example/demo/domain/ingredientDomain/IngredientProjection.java
@@ -1,9 +1,10 @@
 package com.example.demo.domain.ingredientDomain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface IngredientProjection {
     public String getName();
     public UUID getId();
-    public double getPrice();
+    public BigDecimal getPrice();
 }
