+++ src/main/java/com/example/demo/domain/pizzaDomain/PizzaListProjection.java
@@ -1,5 +1,6 @@
 package com.example.demo.domain.pizzaDomain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
@@ -8,7 +9,7 @@ public interface PizzaListProjection {
     public Image getImage();
     public String getName();
     public UUID getId();
-    public double getPrice();
+    public BigDecimal getPrice();
 
     public interface Image{
         public String getPublic_id();
