+++ src/main/java/com/example/demo/dto/ingredientDto/CreateOrUpdateIngredientDTO.java
@@ -1,6 +1,8 @@
 package com.example.demo.dto.ingredientDto;
 
+import java.math.BigDecimal;
+
 public class CreateOrUpdateIngredientDTO {
     public String name;
-    public double price;
+    public BigDecimal price;
 }
