+++ src/main/java/com/example/demo/dto/pizzaDto/PizzaDTO.java
@@ -1,5 +1,6 @@
 package com.example.demo.dto.pizzaDto;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -7,6 +8,6 @@ import java.util.UUID;
 public class PizzaDTO {
     public UUID id;
     public String name;
-    public double price;
+    public BigDecimal price;
     public Set<PizzaIngredientDTO> ingredients = new HashSet<PizzaIngredientDTO>();
 }
